change id type from long to int

+++ backend/Drinker/src/main/java/com/Drinker/repository/AlcoholRepo.java
@@ -3,5 +3,5 @@ package com.Drinker.repository;
 import com.Drinker.model.Alcohol;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlcoholRepo extends JpaRepository<Alcohol, Long> {
+public interface AlcoholRepo extends JpaRepository<Alcohol, Integer> {
 }
